% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSubTrees.R
\name{buildLocalTrees}
\alias{buildLocalTrees}
\title{Build a tree for each window}
\usage{
buildLocalTrees(fileName, DNAwin, contigs, winSize = 1e+05,
  subModel = "raw", minSites, ploidy = 2, nCores = 2, write = NA)
}
\arguments{
\item{fileName}{\code{character}. Full path name for a Tabix indexed VCF file.}

\item{DNAwin}{\code{list} of multiple \code{DNAbin}. A list of already imported VCF files.}

\item{contigs}{\code{character}. Default is \code{"all"} Contigs to extract windows from.}

\item{winSize}{\code{numeric}. Default is \code{100000}. Window size in base pairs.}

\item{subModel}{\code{character}. The substitution model to fit to the data. Accepts nucleotide models defined in \code{ape::dist.dna}:
"raw", "N", "TS", "TV", "JC69", "K80", "F81", "K81", "F84", "BH87", "T92", "TN93", and "GG95."}

\item{minSites}{\code{numeric}. Default is 10 percent of winSize. The minimum number of sites within a window to consider it.}

\item{ploidy}{\code{numeric}. The ploidy of the VCF, as called by Variant Caller.}

\item{nCores}{\code{numeric}. Number of cores to run analysis on.}

\item{write}{\code{character}. Default is NA, will only return the R object. The directory to write the trees out to.}
}
\value{
\code{list} of trees in \code{phy} format. Newick format trees in a specified directory.
}
\description{
Build local trees for each window along a contig.
}
\details{
Builds a tree using data from each window across a contig. Defined as a local tree these
represent the evolutionary history of the window.
}
\examples{

}
