% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSubTrees.R
\name{buildWindowTrees}
\alias{buildWindowTrees}
\title{Three Population Tree Tip Proportion}
\usage{
buildWindowTrees(fileName, contigs, winSize = 1e+05, subModel = "raw",
  minSites, ploidy = 2, output = "tree", nCores = 2)
}
\arguments{
\item{subTrees}{Trees to carry out 3 populaiton Tree Tip proportions on. Can be either a \code{character} file path or \code{multiphy}
object from the package \code{ape}}

\item{metaD}{Trees to carry out 3 populaiton Tree Tip proportions on. Can be either a \code{character} file path or \code{multiphy}
object from the package \code{ape}}
}
\value{
\code{list} of trees in \code{phy} format.
}
\description{
Carries out Three Population Subtree Tip proportions from \code{Ward & Baxter (2018), GBE}
}
\details{
Currently some plots can take a while to render if the \code{FastqcDataList} passed to
\code{fastqcInput} has many elements
}
\examples{

}
