% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyApp.R
\name{popShiny}
\alias{popShiny}
\title{Shiny GUI}
\usage{
popShiny()
}
\arguments{
\item{fileName}{A \code{character} vector of length one containing the full path name for a Tabix indexed VCF}

\item{contigs}{\code{character}. Default is \code{"all"} Contigs to extract windows from.}
}
\value{
Shiny GUI to visulaise and explore results
}
\description{
Shiny GUI
}
\details{
Run the GUI to explore the data generated by this package.
}
\examples{

}
