% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInVcf.R
\name{popStatWindows}
\alias{popStatWindows}
\title{Calculate population statistics along contigs}
\usage{
popStatWindows(fileName, contigs = "all", winSize = 1e+05, minSites,
  ploidy = 2, stat = c("dxy", "pi", "da"), pops, nCores = 1)
}
\arguments{
\item{fileName}{A \code{character} vector of length one containing the full path name for a Tabix indexed VCF}

\item{contigs}{\code{character}. Default is \code{"all"} Contigs to extract windows from.}

\item{winSize}{\code{numeric}. Default is \code{100000}. Window size in base pairs.}

\item{minSites}{\code{numeric}. Default is 10 percent of winSize. The minimum number of sites within a window to consider it.}

\item{ploidy}{\code{numeric}. The ploidy of the VCF, as called by Variant Caller.}

\item{stat}{\code{character}. Default is to perform all statistics. The statistics to carry out on the windows
options are \code{dxy}, \code{pi}, \code{da}, and \code{Dt}.}

\item{pops}{\code{data.frame}. If not suplied statistics will be calculated on an individual basis.
\code{data.frame} should contain two columns \code{sampleNames} specifying sample names as they appear in
the VCF sample field and \code{pop} defines the populations of the samples. See example.}

\item{nCores}{\code{numeric}. Number of cores to run analysis on.}
}
\value{
A \code{tibble} containing populations statistics passed from \code{stat}
}
\description{
Carries out common sequence diversity and distance calculations on individuals or populations.
}
\details{
As reading in the variant data is currently the runtime bottleneck it is advised to run all statistics
simultaniously to reduce overall runtime. \cr
\cr
\code{dxy}: Nei's absolute distance between two populations X and Y. \cr
\code{pi}: Nei's within populaition nucleotide diversity. \cr
\code{da}: Nei's net genetic distance between two populations X and Y \cr
\code{Dt}: Tajimas D neutrality test.
}
\examples{


}
