% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threePopTest.R
\name{threePop}
\alias{threePop}
\title{Three Population Tree Tip Proportion}
\usage{
threePop(x)
}
\arguments{
\item{trees}{Trees to carry out 3 populaiton Tree Tip proportions on. Can be either a \code{character} file path or \code{multiphy}
object from the package \code{ape}
or simply a \code{character} vector of paths to fastqc files.}
}
\value{
UI data for fastQC shiny.
}
\description{
Carries out Three Population Subtree Tip proportions from \code{Ward & Baxter (2018), GBE}
}
\details{
Currently some plots can take a while to render if the \code{FastqcDataList} passed to
\code{fastqcInput} has many elements
}
\examples{
\dontrun{
#' # Get the files included with the package
barcodes <- c("ATTG", "CCGC", "CCGT", "GACC", "TTAT", "TTGG")
suffix <- c("R1_fastqc.zip", "R2_fastqc.zip")
fileList <- paste(rep(barcodes, each = 2), rep(suffix, times = 5), sep = "_")
fileList <- system.file("extdata", fileList, package = "ngsReports")

# Load the FASTQC data as a FastqcDataList
fdl <- getFastqcData(fileList)

# Run the Shiny app
fastqcShiny(fdl)
}

}
