% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeTopologyChecker.R
\name{topologyCheck}
\alias{topologyCheck}
\title{Compare trees to clade structure}
\usage{
topologyCheck(trees, cladeStr)
}
\arguments{
\item{trees}{\code{multiphy} or \code{list} of \code{phy}.
Trees to test for clade structure specified in \code{cladeStr}}

\item{cladeStr}{\code{data.frame}. Clade structure of the tree.}
}
\value{
\code{list} of trees in \code{phy} format containing all discordant trees.
}
\description{
Identifies trees discordant with a tree topology
}
\details{
checks a all trees for a provided clade structure
}
\examples{


}
