% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWindow.R
\name{vcfWindow}
\alias{vcfWindow}
\title{Read window from VCF}
\usage{
vcfWindow(fileName, contig, param, ploidy)
}
\arguments{
\item{fileName}{A \code{character} vector of length one containing the full path name for a Tabix indexed VCF}

\item{contig}{A single element \code{character} vector containing the contig name in the VCF}

\item{param}{An object of class \code{ScanVcfParam} containing a \code{GRanges} object for the target locus}

\item{ploidy}{\code{Integer} or \code{Numeric}. The ploidy of the VCF, as called by Variant Caller}
}
\value{
An object of class \code{DNAbin} from the package \code{ape}
}
\description{
Reads in a specified window from a tabix indexed Variant Call Format
}
\details{
Uses the package \code{VariantAnnotation} to read in a locus (window) from the target VCF,
}
\examples{

}
